fix: order methods
