+++ Qs02-RewardPoints/.env
@@ -1,0 +1,6 @@
+[db]
+host = '127.0.0.1'
+name = 'mycommerce'
+port = '4000'
+user = 'root'
+pass =
