+++ Qs02-RewardPoints/.gitignore
@@ -1,2 +1,2 @@
 vendor
-.env
+# .env
