+++ Qs02-RewardPoints/app/controller/orderController.php
@@ -37,28 +37,30 @@ class OrderController
     public function place(OrderDTO $orderPayload): array
     {
         // any point claim ?
-        $pointClaimed = $this->conversion->convertPointClaimToPointDBRate($orderPayload->currencyId, $orderPayload->pointClaimed);
+        $pointClaimed = $orderPayload->pointClaimed;
 
-        if (!$pointClaimed) {
-            return $this->response->sendResponse('Currency not recognize', 422);
-        }
-
-        // check existing point enough
         if ($pointClaimed > 0) {
+            // any point claim ?
+            $pointConverted = $this->conversion->convertPointClaimToPointDBRate($orderPayload->currencyId, $orderPayload->pointClaimed);
+
+            if (!$pointConverted) {
+                return $this->response->sendResponse('Currency not recognize', 422);
+            }
             $currentPoint = $this->reward->available();
 
-            if ($currentPoint < $pointClaimed) {
+            if ($currentPoint < $pointConverted) {
                 return $this->response->sendResponse('Reward points not enough to be claimed.', 422);
             }
         }
 
         // update status and save order
         $orderPayload->setOrderStatusId(OrderStatus::Pending->value());
+
         $order = $this->order->create($orderPayload);
 
         // record point claim
         if ($pointClaimed > 0) {
-            $this->reward->create($order->userId, TransactionType::Debit, $pointClaimed);
+            $this->reward->create($order->userId, $order->id, TransactionType::Debit->value(), $pointClaimed);
         }
 
         return $this->response->sendResponse($order);
@@ -74,7 +76,11 @@ class OrderController
         }
 
         // update order status
-        $this->order->updateStatus($orderId, 'cancel');
+        $order = $this->order->updateStatus($orderId, 'cancel');
+
+        if (!$order) {
+            return $this->response->sendResponse('Invalid request', 409);
+        }
 
         // reset reward points used
         $this->reward->destroyOrderRecord($orderId);
@@ -93,7 +99,11 @@ class OrderController
         // update order status
         $order = $this->order->updateStatus($orderId, 'paid');
 
-        return $this->response->sendResponse($order, 201);
+        if (!$order) {
+            return $this->response->sendResponse('Invalid request', 409);
+        }
+
+        return $this->response->sendResponse($order, 200);
     }
 
     public function complete(int $orderId)
@@ -107,10 +117,14 @@ class OrderController
         // update order status
         $order = $this->order->updateStatus($orderId, 'complete');
 
+        if (!$order) {
+            return $this->response->sendResponse('Invalid request', 409);
+        }
+
         // give points based on currency
         $rewardPt = $this->conversion->convertPaymentToPointDBRate($order->currencyId, $order->totalSales);
-        $this->reward->create($order->userId, TransactionType::Credit, $rewardPt);
+        $this->reward->create($order->userId, $order->id, TransactionType::Credit->value(), $rewardPt);
 
-        return $this->response->sendResponse($order, 201);
+        return $this->response->sendResponse($order, 200);
     }
 }
